ml_service/requirements.txt
torch
transformers
datasets
scikit-learn
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
joblib
tqdm
matplotlib
